Moved some parsing bits around
