+++ src/packages/fuelrats_api/v3/_dataclasses/users.py
